refactor graph utils to helper
